Add mention of cert-manager cfg needed
I will be referring back to research/ for that. In `wsl-kind-docker.md`
specifically.

NEXT STEPS: debug proxius as is, add some log lines here and there,
create and destroy pods and proxydefs. Figure out how the whole logic
works, then start adding support for automatically creating configmaps
when a proxydef is created, as well as automatically updating a pod spec
when it is created by adding an env from the same configmaps that should
already be created. Simple. Just like SIC.