Struggling to get manifests.yaml to generate properly
but nothing will work.

I'm starting over. And postponing the webhook side of this to later.

First I will start with just having a normal controller to reconcile
ProxyDefs into ConfigMaps. And only after that's verified as working,
will I add a mutating webhook for Pods so that they can inject
environments from the ConfigMap.